+++ bin/www
@@ -3,32 +3,14 @@
 // Express can launch a www executable to handle certain tasks.
 // www goes in bin (sometimes .bin to keep it hidden)
 
-// This way, we can set certain properties here
-// rather than having them take up space in server.js.
-var debug = require('debug')('server');
-
 // we bring in the app and server we exported from server.js
-var serverJS = require('../server');
-var app = serverJS.app;
-var server = serverJS.server;
-
-// we bring in the models we exported with index.js
-// var models = require('../models');
+var app = require('../app');
 
 // we set the port of the app
 app.set('port', process.env.PORT || 3000);
 
-// we sync the models with our db 
-// (thus creating the apropos tables)
-// models.sequelize.sync().then(function() { // {force: true}
-
-	// Create stories since force true deletes everything
-  // require('../config/createStories')();
-
-	// set our app to listen to the port we set above
-  server.listen(app.get('port'), function() {
-  	// then save a log of the listening to our debugger.
-    debug('Express server listening on port ' + server.address().port);
-  });
-
-// });
+// set our app to listen to the port we set above
+var server = app.listen(app.get('port'), function() {
+	// then save a log of the listening to our debugger.
+  console.log('Express server listening on port ' + server.address().port);
+});
