+++ package.json
@@ -21,7 +21,6 @@
     "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.15.2",
     "cookie-parser": "^1.4.3",
-    "debug": "^2.2.0",
     "express": "^4.14.0",
     "express-session": "^1.14.1",
     "mongoose": "^4.6.2",
